Move unwind info module to its own crate
This helps other consumers of this library to not depend on libelf,
zlib, etc that the main crate depends on.

Test Plan
=========

CI
