+++ lightswitch-unwind-info/CHANGELOG.md
@@ -1,0 +1,3 @@
+Unreleased
+----------
+Initial documented release, extracted into its own crate
