+++ lightswitch-unwind-info/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "lightswitch-unwind-info"
+version = "0.0.0"
+edition = "2024"
+description = "Makes unwind information go brrr"
+license = "MIT"
+repository = "https://github.com/javierhonduco/lightswitch"
+
+[dependencies]
+gimli = "0.34.0"
+lazy_static = "1.5.0"
+object = { workspace = true }
+lightswitch-object = { workspace = true }
+plain = { workspace = true }
+memmap2 = { workspace = true }
+anyhow = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true }
+ring = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
